requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
faiss-cpu==1.7.3
sentence-transformers==2.2.2
scikit-learn==1.2.2
nltk==3.8
pydantic==1.10.7
